Added edit log feature to progression tracker

+++ components/Log.js
@@ -1,16 +1,33 @@
-import { StyleSheet, Text, TouchableOpacity, View } from "react-native";
+import { StyleSheet, Text, TouchableOpacity, View, Alert, ShadowPropTypesIOS } from "react-native";
 import { useState } from "react";
 import { progressionLogRemoved } from "../redux/actions.js";
 import store from "../redux/store";
+import Icon from 'react-native-vector-icons/Entypo'
 
 const Log = (props) => {
 
   const onLogRemove = () => {
-    store.dispatch(progressionLogRemoved(props.logInfo.id));
-    props.toggle()
-    console.log("removed");
-    console.log(store.getState())
+    Alert.alert(
+      `Deleting Log Made on ${props.date}` ,
+      "Are you sure you want to delete this log?",
+      [
+        {
+          text: "Delete",
+          onPress: () => {
+            store.dispatch(progressionLogRemoved(props.logInfo.id));
+            props.logUpdate()
+            console.log("Log Deleted")},
+          style: "cancel"
+        },
+        { text: "Cancel", onPress: () => console.log("Canceled") }
+      ]
+    );
   }
+
+  const onLogEdit = () => {
+    props.toggleLogEdit();
+    props.edit(props.logInfo)
+  };
 
   return(
     
@@ -18,6 +35,9 @@ const Log = (props) => {
       <Text style = {{fontSize: 18}}>Progress Log made on {props.date}</Text>
       <TouchableOpacity style = {styles.removeButton} onPress = {() => onLogRemove()}>
         <Text style = {{fontSize: 25, color: "#f00"}}>X</Text>
+      </TouchableOpacity>
+      <TouchableOpacity style = {styles.editButton} onPress={() => onLogEdit()}>
+        <Icon name="edit" size={25} color="#000"/>
       </TouchableOpacity>
     </View>
   )
@@ -38,6 +58,10 @@ const styles = StyleSheet.create({
   removeButton: {
     position: "absolute",
     right: 20 
+  },
+  editButton: {
+    position: "absolute",
+    right: 50 
   }
 })
